# Conflicts:
#	MMVII/include/MMVII_DeclareAllCmd.h
#	MMVII/src/Appli/cSpecMMVII_Appli.cpp
#	MMVII/src/BundleAdjustment/Bundle_BlocRig.cpp
#	MMVII/src/Instrumental/cBlockCamInit.cpp
#	MMVII/src/SymbDerGen/Formulas_BlockRigid.h
#	MMVII/src/SymbDerGen/GenerateCodes.cpp
#	src/uti_phgrm/MICMAC/cEtapeMecComp.cpp

+++ MMVII/src/CodedTarget/cGenerateTarget.cpp
@@ -683,7 +683,7 @@ cCodedTargetPatternIm::tIm cCodedTargetPatternIm::MakeOneImTarget(const cOneEnco
            }
            else if (aLab>=eLPT::eNumB0)
            {
-                bool BitIs_1 =  (aCode & (1<<(int(aLab)-int(eLPT::eNumB0)))) != 0;
+               bool BitIs_1 =  (aCode & (size_t(1)<<(int(aLab)-int(eLPT::eNumB0)))) != 0;
                 isBG = BitIs_1 !=  BGIs_0;
            }
